requirements.txt
streamlit
pymongo
python-dotenv
pandas
google-api-python-client
isodate
plotly
dask[complete]
numpy
scikit-learn
flask